pyproject.toml
[project]
name = "gitdoap"
version = "0.1.0"
description = ""
authors = [
    {name = "Natanael Arndt",email = "arndtn@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0.0"
dependencies = [
    "rdflib (>=7.1.4,<8.0.0)",
    "pygithub (>=2.6.1,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
